Enable assistants to switch to user view
Added to CoursePermissionEvaluator  2 new hasPrivilegedAccessToCourse() methods, which return true for both admins and assistants.
These replace hasAdminAccessToCourse() in the @PreAuthorize decorators of the getCourseStudents functions (see UserService, AdminSubmissionService), allowing assistants to also access the list of students in the course.

To enable the "Switch To User" button, the conditional in impersonateUser was changed to hasPrivilegedAccess.